refactor: rename base_agent.py to atomic_agent.py for v2.0 consistency
  - Renamed base_agent.py to atomic_agent.py to match the new AtomicAgent class name
  - Updated all import statements in __init__.py files
  - Renamed test_base_agent.py to test_atomic_agent.py
  - Aligns file naming with v2.0 class naming conventions (BaseAgent → AtomicAgent)
