+++ atomic-agents/atomic_agents/__init__.py
@@ -3,7 +3,7 @@ Atomic Agents - A modular framework for building AI agents.
 """
 
 # Core exports - base classes only
-from .agents.base_agent import AtomicAgent, AgentConfig, BasicChatInputSchema, BasicChatOutputSchema
+from .agents.atomic_agent import AtomicAgent, AgentConfig, BasicChatInputSchema, BasicChatOutputSchema
 from .base import BaseIOSchema, BaseTool, BaseToolConfig
 
 # Version info
